refactor(email): drive outbox delivery with Hangfire instead of a custom worker
Move email sending out of the API host and into the Cron host, and hand the
delivery machinery (retry back-off, scheduling, crash recovery, dashboard) to
Hangfire rather than the bespoke worker.

- Delete EmailOutboxWorker (polling loop, FOR UPDATE SKIP LOCKED claim, lease)
  and EmailOutboxRetryPolicy (+ test); sending is now a per-email Hangfire job
  EmailOutboxJob.SendAsync with [AutomaticRetry]. Transient failure throws so
  Hangfire retries; the job marks the row Failed on its last attempt (detected
  via PerformContext.RetryCount) so an exhausted row never stays stuck Queued.
- Keep the transactional outbox: the API writes the Pending row in the business
  transaction and publishes the Redis nudge, nothing more.
- Move the whole email stack (providers, templates, dispatcher, options, .liquid
  files) API -> Cron so no ASP.NET/DI logic or MailKit/Fluid lands in Common.
- New coalescing EmailOutboxConsumer in Cron: Redis push + 1-minute safety sweep
  share one wake signal, claim Pending rows, and enqueue Hangfire jobs.
- EmailStatus -> Pending, Queued, Sent, Failed; drop next_attempt_at,
  attempt_started_at, attempt_count (Hangfire owns retry state). Migration
  updated in place; anti-Hangfire docs rewritten for the hybrid design.
