+++ Common/Constants/Constants.cs
@@ -21,17 +21,8 @@ public static class Duration
 
     public static readonly TimeSpan DeactivatedAccountRetentionTime = TimeSpan.FromDays(14);
 
-    /// <summary>How long a claimed email-outbox row may be in flight before it is treated as
-    /// abandoned (worker crashed mid-send) and eligible to be re-claimed.</summary>
-    public static readonly TimeSpan EmailOutboxLeaseTimeout = TimeSpan.FromMinutes(5);
-
-    /// <summary>Safety-net poll interval for the email-outbox consumer. Normal delivery is push-driven
-    /// (Redis notification on enqueue); this only catches due retries and any missed notification.</summary>
-    public static readonly TimeSpan EmailOutboxPollInterval = TimeSpan.FromSeconds(15);
-
-    /// <summary>Base delay for the first email-outbox retry; doubles each subsequent attempt.</summary>
-    public static readonly TimeSpan EmailOutboxRetryBaseDelay = TimeSpan.FromSeconds(30);
-
-    /// <summary>Upper bound on the exponential email-outbox retry back-off.</summary>
-    public static readonly TimeSpan EmailOutboxRetryMaxDelay = TimeSpan.FromHours(2);
+    /// <summary>Safety-net sweep interval for the email-outbox consumer. Normal hand-off is push-driven
+    /// (Redis notification on enqueue); this only catches a missed notification. Hangfire owns retry
+    /// timing for messages already handed off, so this need not be tight.</summary>
+    public static readonly TimeSpan EmailOutboxPollInterval = TimeSpan.FromMinutes(1);
 }
