+++ Cron/Cron.csproj
@@ -5,6 +5,8 @@
 
   <!-- NuGet packages -->
   <ItemGroup>
+    <PackageReference Include="Fluid.Core" />
+    <PackageReference Include="MailKit" />
     <PackageReference Include="Hangfire.AspNetCore" />
     <PackageReference Include="Hangfire.PostgreSql" />
   </ItemGroup>
@@ -14,6 +16,15 @@
     <None Update="devcert.pfx">
       <CopyToPublishDirectory>Always</CopyToPublishDirectory>
     </None>
+  </ItemGroup>
+
+  <!-- Pre-rendered Liquid email templates (committed to repo, not built at compile time).
+       Re-render with `pnpm export` in email-templates/ when you change a .tsx source. -->
+  <ItemGroup>
+    <Content Include="SmtpTemplates\*.liquid">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
+    </Content>
   </ItemGroup>
 
   <!-- Git stuff -->
