+++ Cron/Services/Email/EmailServiceUtils.cs
@@ -1,8 +1,8 @@
 ﻿using System.Net.Mail;
 using MimeKit;
-using OpenShock.API.Services.Email.Mailjet.Mail;
+using OpenShock.Cron.Services.Email.Mailjet.Mail;
 
-namespace OpenShock.API.Services.Email;
+namespace OpenShock.Cron.Services.Email;
 
 public static class EmailServiceUtils
 {
