+++ Cron/Services/Email/NoneEmailService.cs
@@ -1,6 +1,6 @@
-﻿using OpenShock.API.Services.Email.Mailjet.Mail;
+﻿using OpenShock.Cron.Services.Email.Mailjet.Mail;
 
-namespace OpenShock.API.Services.Email;
+namespace OpenShock.Cron.Services.Email;
 
 /// <summary>
 /// This is a noop implementation of the email service. It does nothing.
